requirements.txt
langchain
langchain_community
python-dotenv
ipykernel
langchain_core
streamlit